Dictionary consistency and redundancy changes/updates, E/ to AOE/
Changed most legacy multi-stroke definitions that began with E/ to being with AOE/, in order to be more consistent from a phonetic perspective, to reduce/eliminate conflicts with E/ as brief for "he", per discord conversation May 22, 2018 w/StenoKnight.
Removed redundant multi-stroke definitions that defined the word "he" followed by a second word, wherein the stroke(s) for the second word was identical to existing stroke(s) for the second word, as these are redundant entries.
Removed several duplicate entries.